requirements.txt
PySide6>=6.7.0
pyinstaller
Pillow